LPS-200563 Inline UserGroupMembershipPolicyFactoryImpl logic into Util
