+++ portal-impl/src/com/liferay/portal/security/membershippolicy/UserGroupMembershipPolicyFactoryUtil.java
@@ -5,8 +5,18 @@
 
 package com.liferay.portal.security.membershippolicy;
 
+import com.liferay.petra.concurrent.DCLSingleton;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.module.util.SystemBundleUtil;
 import com.liferay.portal.kernel.security.membershippolicy.UserGroupMembershipPolicy;
-import com.liferay.portal.kernel.security.membershippolicy.UserGroupMembershipPolicyFactory;
+import com.liferay.portal.util.PropsValues;
+
+import org.osgi.framework.BundleContext;
+import org.osgi.framework.ServiceReference;
+import org.osgi.util.tracker.ServiceTracker;
+import org.osgi.util.tracker.ServiceTrackerCustomizer;
 
 /**
  * @author Roberto Díaz
@@ -14,16 +24,73 @@ import com.liferay.portal.kernel.security.membershippolicy.UserGroupMembershipPo
 public class UserGroupMembershipPolicyFactoryUtil {
 
 	public static UserGroupMembershipPolicy getUserGroupMembershipPolicy() {
-		return _userGroupMembershipPolicyFactory.getUserGroupMembershipPolicy();
+		ServiceTracker<UserGroupMembershipPolicy, UserGroupMembershipPolicy>
+			serviceTracker = _serviceTrackerDCLSingleton.getSingleton(
+				UserGroupMembershipPolicyFactoryUtil::_createServiceTracker);
+
+		return serviceTracker.getService();
 	}
 
-	public void setUserGroupMembershipPolicyFactory(
-		UserGroupMembershipPolicyFactory userGroupMembershipPolicyFactory) {
+	private static ServiceTracker
+		<UserGroupMembershipPolicy, UserGroupMembershipPolicy>
+			_createServiceTracker() {
 
-		_userGroupMembershipPolicyFactory = userGroupMembershipPolicyFactory;
+		ServiceTracker<UserGroupMembershipPolicy, UserGroupMembershipPolicy>
+			serviceTracker = new ServiceTracker<>(
+				_bundleContext, UserGroupMembershipPolicy.class,
+				new UserGroupMembershipPolicyTrackerCustomizer());
+
+		serviceTracker.open();
+
+		return serviceTracker;
 	}
 
-	private static UserGroupMembershipPolicyFactory
-		_userGroupMembershipPolicyFactory;
+	private static final Log _log = LogFactoryUtil.getLog(
+		UserGroupMembershipPolicyFactoryUtil.class);
+
+	private static final BundleContext _bundleContext =
+		SystemBundleUtil.getBundleContext();
+	private static final DCLSingleton
+		<ServiceTracker<UserGroupMembershipPolicy, UserGroupMembershipPolicy>>
+			_serviceTrackerDCLSingleton = new DCLSingleton<>();
+
+	private static class UserGroupMembershipPolicyTrackerCustomizer
+		implements ServiceTrackerCustomizer
+			<UserGroupMembershipPolicy, UserGroupMembershipPolicy> {
+
+		@Override
+		public UserGroupMembershipPolicy addingService(
+			ServiceReference<UserGroupMembershipPolicy> serviceReference) {
+
+			UserGroupMembershipPolicy userGroupMembershipPolicy =
+				_bundleContext.getService(serviceReference);
+
+			if (PropsValues.MEMBERSHIP_POLICY_AUTO_VERIFY) {
+				try {
+					userGroupMembershipPolicy.verifyPolicy();
+				}
+				catch (PortalException portalException) {
+					_log.error(portalException);
+				}
+			}
+
+			return userGroupMembershipPolicy;
+		}
+
+		@Override
+		public void modifiedService(
+			ServiceReference<UserGroupMembershipPolicy> serviceReference,
+			UserGroupMembershipPolicy userGroupMembershipPolicy) {
+		}
+
+		@Override
+		public void removedService(
+			ServiceReference<UserGroupMembershipPolicy> serviceReference,
+			UserGroupMembershipPolicy userGroupMembershipPolicy) {
+
+			_bundleContext.ungetService(serviceReference);
+		}
+
+	}
 
 }
